# libs/ts/identity-sdk-core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nvbes Identity API",
    "description": "Authentication, authorization, and identity management API",
    "contact": {
      "name": "nvbes",
      "url": "https://nvbes.fr"
    },
    "license": {
      "name": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://identity.nvbes.fr",
      "description": "Production"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "list_workspaces",
        "responses": {
          "200": {
            "description": "List workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workspace created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspaceId}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "workspaces"
        ],
        "operationId": "update_workspace",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Workspace updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "UpdateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
      "WorkspaceListResponse": {
        "type": "object",
        "required": [
          "workspaces"
        ],
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkspaceView"
            }
          }
        }
      },
      "WorkspacePolicyView": {
        "type": "object",
        "required": [
          "member_can_create_share_links",
          "require_admin_approval_for_member_share",
          "default_share_link_ttl_days",
          "max_share_link_ttl_days"
        ],
        "properties": {
          "default_share_link_ttl_days": {
            "type": "integer",
            "format": "int32"
          },
          "max_share_link_ttl_days": {
            "type": "integer",
            "format": "int32"
          },
          "member_can_create_share_links": {
            "type": "boolean"
          },
          "require_admin_approval_for_member_share": {
            "type": "boolean"
          }
        }
      },
      "WorkspaceResponse": {
        "type": "object",
        "required": [
          "workspace"
        ],
        "properties": {
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceView"
          }
        }
      },
...
      "WorkspaceView": {
        "type": "object",
        "required": [
          "id",
          "owner_principal_id",
          "name",
          "workspace_type",
          "data_region",
          "jurisdiction",
          "role",
          "plan_code",
          "policy",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "data_region": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "jurisdiction": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "owner_principal_id": {
            "type": "string",
            "format": "uuid"
          },
          "plan_code": {
            "type": "string"
          },
          "policy": {
            "$ref": "#/components/schemas/WorkspacePolicyView"
          },
          "role": {
            "type": "string"
          },
          "trial_ends_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "workspace_type": {
            "type": "string"
          }
        }
      }
    }
  },
...
}